+++ src/compiler/sema_liveness.c
@@ -561,7 +561,6 @@ RETRY:
 		case DECL_CT_ASSERT:
 		case DECL_CT_ECHO:
 		case DECL_CT_EXEC:
-		case DECL_CT_EXPAND:
 		case DECL_IMPORT:
 		case DECL_CT_INCLUDE:
 		case DECL_LABEL: