+++ src/compiler/symtab.c
@@ -314,6 +314,7 @@ void symtab_init(uint32_t capacity)
 	attribute_list[ATTRIBUTE_BUILTIN] = KW_DEF("@builtin");
 	attribute_list[ATTRIBUTE_CALLCONV] = KW_DEF("@callconv");
 	attribute_list[ATTRIBUTE_COMPACT] = KW_DEF("@compact");
+	attribute_list[ATTRIBUTE_CONST] = KW_DEF("@const");
 	attribute_list[ATTRIBUTE_DEPRECATED] = KW_DEF("@deprecated");
 	attribute_list[ATTRIBUTE_DYNAMIC] = KW_DEF("@dynamic");
 	attribute_list[ATTRIBUTE_EXPORT] = KW_DEF("@export");