+++ src/compiler/tokens.c
@@ -362,8 +362,6 @@ const char *token_type_to_string(TokenType type)
 			return "$error";
 		case TOKEN_CT_EXEC:
 			return "$exec";
-		case TOKEN_CT_EXPAND:
-			return "$expand";
 		case TOKEN_CT_EXTNAMEOF:
 			return "$extnameof";
 		case TOKEN_CT_FEATURE: